todo finalizado

+++ src/dao/mongo/usersMongo.js
@@ -1,5 +1,6 @@
 import { logger } from "../../utils/logger.js";
 import UserModel from "../mongo/models/users.model.js";
+import { sendMail } from "../../utils/transport.js";
 
 export const getAllUsers = async (req, res) => {
   return await UserModel.find();
@@ -23,15 +24,32 @@ export const deleteUser = async (uid) => {
 
 export const deleteInactiveUsers = async () => {
   try {
-    const twoDaysMilliseconds = 2 * 24 * 60 * 60 * 1000;
-    const threshold = new Date(Date.now() - twoDaysMilliseconds);
-
-    const deletionResult = await UserModel.deleteMany({
+    const fiveMinutesMilliseconds = 2 * 24 * 60 * 60 * 1000;
+    const threshold = new Date(Date.now() - fiveMinutesMilliseconds);
+    const inactiveUsers = await UserModel.find({
       last_connection: { $lt: threshold },
     });
 
-    if (deletionResult.deletedCount > 0) {
+    if (inactiveUsers.length > 0) {
+      inactiveUsers.forEach(async (user) => {
+        const email = user.email;
+        sendMail(
+          email,
+          "ACCOUNT DELETED",
+          `
+          <h1 style="background:#93c5fd;text-align:center">CODEDOM</h1>
+          <p>We are sending this email to inform you that your account has been deleted due to inactivity. 
+          You can always register again <a href="http://127.0.0.1:8080/register" class="italic">clicking here</a>!</p>
+          `
+        );
+      });
+
+      const deletionResult = await UserModel.deleteMany({
+        last_connection: { $lt: threshold },
+      });
+
       logger.info(`${deletionResult.deletedCount} inactive users deleted.`);
+
       return {
         status: "success",
         deletedCount: deletionResult.deletedCount,
